backend/pyproject.toml
[tool.poetry]
name = "doglog-backend"
version = "0.1.0"
description = "A webapp for tracking dog training"
authors = ["Ruhi Bloodworth <ruhi.bloodworth@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
sqlalchemy = "^2.0.7"
argon2-cffi = "^21.3.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
